+++ CHANGELOG.md
@@ -16,6 +16,7 @@ https://keepachangelog.com/en/1.0.0/
 ### Added
 
 - Makefile rule to build executable with PyInstaller
+- New field `report_descriptor` to `DeviceInfo`
 
 ### Changed